implement widget split
