+++ package.json
@@ -2,6 +2,10 @@
   "name": "@tinystacks/ops-core-widgets",
   "version": "0.0.15",
   "main": "dist/index.js",
+  "exports": {
+    "node": "./dist/node.js",
+    "browser": "./dist/browser.js"
+  },
   "type": "module",
   "files": [
     "dist"
@@ -37,8 +41,9 @@
     "@jest/globals": "^29.5.0",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^14.0.0",
-    "@tinystacks/ops-core": "^0.3.0",
+    "@tinystacks/ops-core": "file:../ops-core/tinystacks-ops-core-0.3.2.tgz",
     "@tinystacks/ops-model": "^0.4.0",
+    "@types/http-status-codes": "^1.2.0",
     "@types/jest": "^29.4.0",
     "@types/lodash.get": "^4.4.7",
     "@types/lodash.isempty": "^4.4.7",
@@ -72,8 +77,9 @@
   },
   "peerDependencies": {
     "@chakra-ui/react": "2.x",
-    "@tinystacks/ops-core": "0.x",
+    "@tinystacks/ops-core": "file:../ops-core/tinystacks-ops-core-0.3.2.tgz",
     "@tinystacks/ops-model": "0.x",
+    "http-status-codes": "2.x",
     "lodash.get": "4.x",
     "lodash.isempty": "4.x",
     "lodash.isnil": "4.x",
