+++ src/node.ts
@@ -1,5 +1,4 @@
 export * as Controllers from './controllers/index.js';
 export * as Models from './models/index.js';
-export * as Views from './views/index.js';
 export * as OpsTypes from './ops-types.js';
 export * from './utils/index.js';
