+++ src/views/components/index.ts
@@ -1,0 +1,1 @@
+export * from './key-value-stat.js';
